refactor matches
