+++ src/pages/landing.js
@@ -1,19 +1,10 @@
 import * as React from 'react';
 import { useState, useEffect } from 'react';
-import { styled } from '@mui/material/styles';
 import Box from '@mui/material/Box';
-import Paper from '@mui/material/Paper';
 import Grid from '@mui/material/Grid';
 
 import { getMatches } from '../services/football';
-
-const Item = styled(Paper)(({ theme }) => ({
-  ...theme.typography.body2,
-  padding: theme.spacing(1),
-  textAlign: 'center',
-  color: theme.palette.text.secondary,
-}));
-
+import Match from '../components/match';
 
 export default function Matches() {
     const [matches, setMatches] = useState([]);
@@ -34,11 +25,7 @@ export default function Matches() {
         <Grid container spacing={3}>
             {matches.map((match) => {
                 return (
-                    <Grid key={match.id} item xs="auto">
-                        <Item>Home: {match.homeTeam.name}</Item>
-                        <Item>Away: {match.awayTeam.name}</Item>
-                        <Item>{match.score.fullTime.homeTeam} - {match.score.fullTime.awayTeam}</Item>
-                    </Grid>
+                    <Match key={match.id} match={match}/>
                 );
             })}
         </Grid>
